// services/graph-rewriting-service/src/types/graph.schema.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface GraphSchema {
  options: {
    /**
     * Type of the graph: either directed or undirected
     */
    type: "directed" | "undirected";
  };
  nodes: GraphNodeSchema[];
  edges: GraphEdgeSchema[];
}
export interface GraphNodeSchema {
  /**
   * The node's ID, also used as node in an edges source/target etc.
   */
  key: string;
  /**
   * The node's attributes.
   */
  attributes: {
    [k: string]: number | string | boolean;
  };
}
export interface GraphEdgeSchema {
  /**
   * The edge's ID
   */
  key: string;
  /**
   * The key of the node at the edge's source
   */
  source: string;
  /**
   * The key of the node at the edge's target
   */
  target: string;
  /**
   * The edges attributes & values
   */
  attributes: {
    [k: string]: number | string | boolean;
  };
}